Fix logs, update types, refactor to form view
